feat: adjusts delete brand flow

+++ src/view/pages/Products/components/BrandsTab/components/DeleteBrandAction.tsx
@@ -1,5 +1,15 @@
+import { useState } from 'react'
+import { Loader2 } from 'lucide-react'
+
 import { Button } from '@/view/components/ui/button'
-import { Alert } from '@/view/components/Alert'
+import {
+  AlertDialog,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/view/components/ui/alert-dialog'
 
 interface DeleteBrandActionProps {
   isOpen: boolean
@@ -8,26 +18,38 @@ interface DeleteBrandActionProps {
 }
 
 export const DeleteBrandAction = ({ isOpen, onDelete, onClose }: DeleteBrandActionProps) => {
+  const [isLoading, setIsLoading] = useState(false)
+
   return (
-    <Alert
-      isOpen={isOpen}
-      title="Apagar marca"
-      description={`
-        Deseja mesmo apagar esta marca?
-        Ao remover a marca, os produtos associados a ela deixarão de ter uma marca associada.
-      `}
-      cancelButton={<Button variant="outline">Cancelar</Button>}
-      proceedButton={
-        <Button
-          variant="destructive"
-          onClick={async () => {
-            await onDelete()
-          }}
-        >
-          Apagar
-        </Button>
-      }
-      onClose={onClose}
-    />
+    <AlertDialog open={isOpen}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Apagar marca</AlertDialogTitle>
+          <AlertDialogDescription>
+            Deseja mesmo apagar esta marca? Ao remover a marca, os produtos associados a ela deixarão de ter uma marca
+            associada.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <Button variant="outline" disabled={isLoading} onClick={onClose}>
+            Cancelar
+          </Button>
+
+          <Button
+            variant="destructive"
+            onClick={async () => {
+              setIsLoading(true)
+              await onDelete()
+              onClose()
+              setIsLoading(false)
+            }}
+            disabled={isLoading}
+          >
+            {isLoading && <Loader2 className="mr-2 w-4 h-4 animate-spin" />}
+            Apagar
+          </Button>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   )
 }
